build(Icons): convert to ts, use tsx, restructure files
