+++ packages/icons/src/utils/converter/index.ts
@@ -1,0 +1,3 @@
+export * from "./formatSVG";
+export * from "./generateComponent";
+export * from "./removeStyle";
