+++ packages/icons/src/utils/converter/removeStyle.ts
@@ -1,10 +1,8 @@
 /**
  * Removes style tags from a node and its children
- * *Note* This mututates the passed in node
- * @param  node A node produced by jsdom
- * @return undefined
+ * *Note* This mutates the passed in node
  */
-const removeStyle = (node) => {
+export const removeStyle = (node: HTMLElement) => {
   const elements = node.getElementsByTagName("*");
   Object.keys(elements).forEach((key) => {
     if (elements[key].removeAttribute) {
@@ -12,5 +10,3 @@ const removeStyle = (node) => {
     }
   });
 };
-
-export default removeStyle;
